refactor: Tcp client 베이스 클래스로 추출
