+++ common/tcp/NodeType.h
@@ -1,0 +1,14 @@
+#ifndef NODE_TYPE_H
+#define NODE_TYPE_H
+
+#include <string>
+
+namespace tcp
+{
+	typedef struct host {
+		std::string ip;
+		int port;
+	} host_t;
+}
+
+#endif // NODE_TYPE_H
