+++ test/fixture/Fixture.cpp
@@ -2,9 +2,8 @@
 
 namespace fixture
 {
-    cctv::host_t cctv1 = { 
+    tcp::host_t cctv1 = { 
         .ip = "192.168.50.14",
-        // .ip = "localhost",
 		.port = 12345
     };
 }
